pyproject.toml
[project]
name = "microservices-final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3==1.28.0",
    "moto>=4.1.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
]
